Refactor
Refactoring each Branch
