--- lib/ui/movie_list_page.txt
@@ -1,65 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MovieListPage extends StatelessWidget {
-  static String tag = 'movie-page';
-  const MovieListPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text("Daftar Film"),
-      ),
-      body: ListView(
-        children: [
-          Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  Column(
-                    children: [
-                      Row(
-                        children: [
-                          Column(
-                            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                            children: [
-                              Row(
-                                children: [
-                                  SizedBox(
-                                    height: 260,
-                                    width: 160,
-                                    child: Image.asset("assets/indigo.jpg", fit: BoxFit.fill,),
-                                  ),
-                                  
-                                ],  
-                              ),
-                              Row(
-                                children: [
-                                  const Text("Indigo"),
-                              ],
-                              )
-                            ],
-                          ),
-                          Column(
-                            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                            children: [
-                              SizedBox(
-                                    height: 260,
-                                    width: 160,
-                                    child: Image.asset("assets/indigo.jpg", fit: BoxFit.fill,),
-                                  ),
-                              const Text("Indigo"),
-                            ],
-                          )
-                        ],
-                      )
-                    ],
-                  )
-                ],
-                )
-              ),
-        ],
-      )
-    );
-  }
-}
